% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_variant_heatmap_tree}
\alias{create_variant_heatmap_tree}
\title{Create a tree visualization with a variant heatmap}
\usage{
create_variant_heatmap_tree(tree, data, output_file)
}
\arguments{
\item{tree}{A phylo object to visualize}

\item{data}{Dataframe with variant data to display as a heatmap}

\item{output_file}{Optional file path to save the plot}
}
\value{
A ggtree plot object
}
\description{
This function generates a publication-quality tree visualization that includes
a heatmap showing variant states (REF/ALT) for each taxon. The visualization
can be customized with various parameters.
}
