% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{flip_tree}
\alias{flip_tree}
\title{Reverse the tree plotting order by reversing edges}
\usage{
flip_tree(tree)
}
\arguments{
\item{tree}{A phylo object to be flipped}
}
\value{
A phylo object with reversed edge order
}
\description{
Reverse the tree plotting order by reversing edges
}
