% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_functions.R
\name{get_variant_gt}
\alias{get_variant_gt}
\title{This function locates a specific variant in a sequence alignment and extracts
the nucleotide at that position for each sequence. It uses a pattern matching
approach to identify the exact position of the variant.}
\usage{
get_variant_gt(data, alignment, reference_seq_index = 1)
}
\arguments{
\item{data}{A data.frame of variant names and  patterns to find}

\item{alignment}{A DNAStringSet object containing the aligned sequences}

\item{reference_seq_index}{Integer index of the reference sequence (default: 1)}
}
\value{
A factor with the nucleotide at the variant position for each sequence
}
\description{
This function locates a specific variant in a sequence alignment and extracts
the nucleotide at that position for each sequence. It uses a pattern matching
approach to identify the exact position of the variant.
}
